omaha_response_handler_action: skip kernels on really old systems
The bootloader configuration in CoreOS <= 522.x.x was SYSLINUX based and
used different install paths. If update_engine is running on such a
system it shouldn't install kernels to the modern location since that
would be wasting space in a somewhat small partition.

The post install script can remain as-is to support these systems,
update_engine can just ignore the kernel.